fix: route delete trip

+++ src/routes/delete-trip.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod'
+import { env } from '../env';
 import { prisma } from "../lib/prisma";
 import { FastifyInstance } from "fastify";
 import { ClientError } from '../errors/client-error';
@@ -11,21 +12,32 @@ export async function deleteTrip(app: FastifyInstance){
         tripId: z.string().uuid(),
       })
     }
-  }, async (request) => {
+  }, async (request, reply) => {
     const { tripId } = request.params
 
     const trip = await prisma.trip.findUnique({
-      where: { id: tripId }
+      where: { id: tripId },
+      include: {
+        participants: {
+          where: {
+            is_owner: false
+          }
+        }
+      }
     })
 
     if(!trip){
       throw new ClientError('Trip not found')
     }
 
+    if(trip.is_confirmed){
+      throw new ClientError('Trip already confirmed')
+    }
+
     await prisma.trip.delete({
       where: { id: tripId },
     })
 
-    return { tripId: trip.id }
+    return reply.redirect(`${env.WEB_BASE_URL}`)
   })
 }
